# internal/generated/api/types.gen.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.1 DO NOT EDIT.
package api

import (
	"time"
)

// Defines values for ErrorCode.
const (
	INTERNALERR   ErrorCode = "INTERNAL_ERR"
	NOCANDIDATE   ErrorCode = "NO_CANDIDATE"
	NOTASSIGNED   ErrorCode = "NOT_ASSIGNED"
	NOTFOUND      ErrorCode = "NOT_FOUND"
	NOTINTEAM     ErrorCode = "NOT_IN_TEAM"
	PREXISTS      ErrorCode = "PR_EXISTS"
	PRMERGED      ErrorCode = "PR_MERGED"
	TEAMEXISTS    ErrorCode = "TEAM_EXISTS"
	VALIDATIONERR ErrorCode = "VALIDATION_ERR"
)

// Defines values for PullRequestStatus.
const (
	MERGED PullRequestStatus = "MERGED"
	OPEN   PullRequestStatus = "OPEN"
)

// CreatePullRequestResponse defines model for CreatePullRequestResponse.
type CreatePullRequestResponse struct {
	Pr PullRequest `json:"pr"`
}

// Error defines model for Error.
type Error struct {
	Code    ErrorCode `json:"code"`
	Message string    `json:"message"`
}

// ErrorCode defines model for Error.Code.
type ErrorCode string

// ErrorResponse defines model for ErrorResponse.
type ErrorResponse struct {
	Error Error `json:"error"`
}

// MergePullRequestResponse defines model for MergePullRequestResponse.
type MergePullRequestResponse struct {
	Pr PullRequest `json:"pr"`
}

// PullRequest defines model for PullRequest.
type PullRequest struct {
	// AssignedReviewers user_id назначенных ревьюверов (0..2)
	AssignedReviewers []string    `json:"assigned_reviewers"`
	AuthorId          string      `json:"author_id"`
	CreatedAt         *time.Time  `json:"createdAt"`
	MergedAt          *time.Time  `json:"mergedAt"`
	PullRequestId     string      `json:"pull_request_id"`
	PullRequestName   string      `json:"pull_request_name"`
	Status            interface{} `json:"status"`
}

// PullRequestShort defines model for PullRequestShort.
type PullRequestShort struct {
	AuthorId        string            `json:"author_id"`
	PullRequestId   string            `json:"pull_request_id"`
	PullRequestName string            `json:"pull_request_name"`
	Status          PullRequestStatus `json:"status"`
}

// PullRequestStatus defines model for PullRequestStatus.
type PullRequestStatus string

// PullRequestsStats defines model for PullRequestsStats.
type PullRequestsStats struct {
	AssignmentsCount int    `json:"assignments_count"`
	PullRequestId    string `json:"pull_request_id"`
}

// ReassignPullRequestResponse defines model for ReassignPullRequestResponse.
type ReassignPullRequestResponse struct {
	Pr PullRequest `json:"pr"`

	// ReplacedBy user_id нового ревьювера
	ReplacedBy string `json:"replaced_by"`
}

// SetIsActiveResponse defines model for SetIsActiveResponse.
type SetIsActiveResponse struct {
	User User `json:"user"`
}

// Stats defines model for Stats.
type Stats struct {
	PullRequestsStats []PullRequestsStats `json:"pull_requests_stats"`
	UserStats         []UserStats         `json:"user_stats"`
}

// Team defines model for Team.
type Team struct {
	Members  []TeamMember `json:"members"`
	TeamName string       `json:"team_name"`
}

// TeamAddResponse defines model for TeamAddResponse.
type TeamAddResponse struct {
	Team Team `json:"team"`
}

// TeamMember defines model for TeamMember.
type TeamMember struct {
	IsActive bool   `json:"is_active"`
	UserId   string `json:"user_id"`
	Username string `json:"username"`
}

// User defines model for User.
type User struct {
	IsActive bool   `json:"is_active"`
	TeamName string `json:"team_name"`
	UserId   string `json:"user_id"`
	Username string `json:"username"`
}

// UserStats defines model for UserStats.
type UserStats struct {
	AssignmentsCount   int    `json:"assignments_count"`
	StatusChangesCount int    `json:"status_changes_count"`
	UserId             string `json:"user_id"`
}

// UsersGetReviewResponse defines model for UsersGetReviewResponse.
type UsersGetReviewResponse struct {
	PullRequests []PullRequestShort `json:"pull_requests"`
	UserId       string             `json:"user_id"`
}

// TeamNameQuery defines model for TeamNameQuery.
type TeamNameQuery = string

// UserIdQuery defines model for UserIdQuery.
type UserIdQuery = string

// PostPullRequestCreateJSONBody defines parameters for PostPullRequestCreate.
type PostPullRequestCreateJSONBody struct {
	AuthorId        string `json:"author_id"`
	PullRequestId   string `json:"pull_request_id"`
	PullRequestName string `json:"pull_request_name"`
}

// PostPullRequestMergeJSONBody defines parameters for PostPullRequestMerge.
type PostPullRequestMergeJSONBody struct {
	PullRequestId string `json:"pull_request_id"`
}

// PostPullRequestReassignJSONBody defines parameters for PostPullRequestReassign.
type PostPullRequestReassignJSONBody struct {
	OldUserId     string `json:"old_user_id"`
	PullRequestId string `json:"pull_request_id"`
}

// GetTeamGetParams defines parameters for GetTeamGet.
type GetTeamGetParams struct {
	// TeamName Уникальное имя команды
	TeamName TeamNameQuery `form:"team_name" json:"team_name"`
}

// GetUsersGetReviewParams defines parameters for GetUsersGetReview.
type GetUsersGetReviewParams struct {
	// UserId Идентификатор пользователя
	UserId UserIdQuery `form:"user_id" json:"user_id"`
}

// PostUsersSetIsActiveJSONBody defines parameters for PostUsersSetIsActive.
type PostUsersSetIsActiveJSONBody struct {
	IsActive bool   `json:"is_active"`
	UserId   string `json:"user_id"`
}

// PostPullRequestCreateJSONRequestBody defines body for PostPullRequestCreate for application/json ContentType.
type PostPullRequestCreateJSONRequestBody PostPullRequestCreateJSONBody

// PostPullRequestMergeJSONRequestBody defines body for PostPullRequestMerge for application/json ContentType.
type PostPullRequestMergeJSONRequestBody PostPullRequestMergeJSONBody

// PostPullRequestReassignJSONRequestBody defines body for PostPullRequestReassign for application/json ContentType.
type PostPullRequestReassignJSONRequestBody PostPullRequestReassignJSONBody

// PostTeamAddJSONRequestBody defines body for PostTeamAdd for application/json ContentType.
type PostTeamAddJSONRequestBody = Team

// PostUsersSetIsActiveJSONRequestBody defines body for PostUsersSetIsActive for application/json ContentType.
type PostUsersSetIsActiveJSONRequestBody PostUsersSetIsActiveJSONBody